apps/studio/deployments/vm-caddy/junjo_app/requirements.txt
junjo==0.61.1
python-dotenv
loguru